pyproject.toml
[tool.poetry]
name = "keensms2mqtt"
version = "0.1.0"
description = ""
authors = ["Leonid Amirov <leonid.amirov@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
aiomqtt = "^2.3.0"
pyyaml = "^6.0.2"
simple-keenetic-client = "^0.1.2"

[tool.poetry.group.dev.dependencies]
ipython = "^8.29.0"
black = "^24.10.0"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
